+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+* Working with datetime conditions in case of non-indexed fields or
+  non-iterating indexes (#373).
+
 ## [1.4.3] - 05-02-24
 
 ### Fixed